Forced int64 in labtrans

+++ pycox/preprocessing/label_transforms.py
@@ -133,7 +133,7 @@ class LabTransDiscreteTime:
         durations = durations.astype(self._dtype)
         events = _values_if_series(events)
         idx_durations, events = self.idu.transform(durations, events)
-        return idx_durations, events.astype('float32')
+        return idx_durations.astype('int64'), events.astype('float32')
 
     @property
     def out_features(self):
